==> notebooks/02_data_summary.ipynb <==
import pandas as pd

# 모든 열이 출력되도록 설정
pd.set_option('display.max_columns', None)

df = pd.read_csv("../data/processed/db_network_data_table.csv")

# 12.3s

### 데이터 요약

df.info()

# 0.0s

df.describe()

# 4.3s

df.head(5)

# 0.0s

### 결측치 확인

# 전체 데이터에 결측치가 있는지 확인
df.isnull().values.any()

# 0.2s

### 데이터의 음수 값 확인

# 각 칼럼에서 음수인 값 개수 확인
df.lt(0).sum()

# 0.1s

### 비슷한 칼럼이 동일한지 확인

# 두 칼럼이 동일한지 확인
is_equal = df['Fwd Header Length'] == df['Fwd Header Length.1']

# 동일한 값이 있는지 여부 출력
if is_equal.all():
    print("두 칼럼 'Fwd Header Length'와 'Fwd Header Length.1'은 동일합니다.")
else:
    print("두 칼럼 'Fwd Header Length'와 'Fwd Header Length.1'은 동일하지 않습니다.")
    
    # 동일하지 않은 인덱스 출력
    print("동일하지 않은 인덱스:", is_equal[is_equal == False].index)

# 0.0s
