==> notebooks/04_features_analysis.ipynb <==
import pandas as pd
import numpy as np
# 모든 열이 출력되도록 설정
pd.set_option('display.max_columns', None)

df = pd.read_csv("../data/processed/db_network_data_table.csv")
df = df.drop(columns=['id', 'source'])
# 12.0s

# 상관계수 계산
correlation_matrix = df.corr()

# 상삼각형 인덱스만 가져오기 (k=1: 자기 자신을 제외한 값들만 선택)
upper_triangle = np.triu_indices_from(correlation_matrix, k=1)

# 상관계수 절대값이 0.8 이상인 변수들만 추출
threshold = 0.8
high_corr_pairs = []  # 상관계수 0.8 이상인 변수 쌍을 저장할 리스트

# 상관계수가 0.8 이상인 변수 쌍을 찾아서 리스트에 추가
for i, j in zip(upper_triangle[0], upper_triangle[1]):
    corr_value = correlation_matrix.iloc[i, j]
    
    # 상관계수가 절대값으로 0.8 이상이면 변수 쌍을 리스트에 추가
    if abs(corr_value) >= threshold:
        high_corr_pairs.append((correlation_matrix.columns[i], correlation_matrix.columns[j], corr_value))

# 상관관계가 높은 변수 쌍 출력
print(f"상관계수가 {threshold} 이상인 변수 쌍: {len(high_corr_pairs)}")
for pair in high_corr_pairs:
    print(f"{pair[0]} - {pair[1]}: {pair[2]:.2f}")

# 다중공선성 파악을 위한 변수 선택
if high_corr_pairs:
    # 상관계수가 0.8 이상인 변수들만 추출하여 리스트로 만듦
    correlated_columns = list({x[0] for x in high_corr_pairs} | {x[1] for x in high_corr_pairs})
    print("\n다중공선성 가능성이 높은 변수들:")
    print(correlated_columns)
else:
    print("다중공선성 문제는 발견되지 않았습니다.")

# 26.2s

# 단 하나의 값으로만 이루어진 칼럼 찾기
single_value_columns = [col for col in df.columns if df[col].nunique() == 1]
df_cleaned = df.drop(columns=single_value_columns)

# 결과 출력
print("제거된 단 하나의 값으로만 이루어진 칼럼:")
print(single_value_columns)

# 2.5s

# 각 칼럼의 값을 tuple로 변환하여 비교
grouped_columns = {}

for col in df_cleaned.columns:
    col_values = tuple(df_cleaned[col].values)  # 각 칼럼의 값을 tuple로 변환
    if col_values in grouped_columns:  # 동일한 값이 이미 존재하는지 확인
        grouped_columns[col_values].append(col)  # 그룹에 추가
    else:
        grouped_columns[col_values] = [col]  # 새로운 그룹 생성

# 그룹화된 칼럼 중 2개 이상 포함된 것만 추출
grouped_result = [group for group in grouped_columns.values() if len(group) > 1]

# 결과 출력
print("그룹화된 칼럼들:")
for group in grouped_result:
    print(group)

# 14.4s

# 그룹화된 칼럼에서 첫 번째 칼럼을 제외한 나머지 칼럼들을 columns_to_drop에 추가
columns_to_drop = [col for group in grouped_result for col in group[1:]]

# df_cleaned에서 동일한 칼럼 삭제
df_cleaned = df_cleaned.drop(columns=columns_to_drop)

# 결과 출력
print("제거된 중복 칼럼:")
print(columns_to_drop)

# 2.6s

# 상관계수 행렬 계산
correlation_matrix = df_cleaned.corr()

# 상삼각형 인덱스만 가져오기 (k=1: 자기 자신을 제외한 값들만 선택)
upper_triangle = np.triu_indices_from(correlation_matrix, k=1)

# 상관계수 절대값이 0.9 이상인 변수들만 추출
threshold = 0.9
columns_to_drop = []  # 삭제할 변수들을 저장할 리스트
high_corr_pairs = []  # 상관계수 0.9 이상인 변수 쌍을 저장할 리스트

# 상관계수가 0.9 이상인 변수 쌍을 찾아서 리스트에 추가
for i, j in zip(upper_triangle[0], upper_triangle[1]):
    corr_value = correlation_matrix.iloc[i, j]
    
    # 상관계수가 절대값으로 0.9 이상이면
    if abs(corr_value) >= threshold:
        columns_to_drop.append(correlation_matrix.columns[j])  # 두 번째 변수만 추가 (첫 번째 변수는 남김)
        high_corr_pairs.append((correlation_matrix.columns[i], correlation_matrix.columns[j], corr_value))  # 변수 쌍과 상관계수 저장

# 중복 제거 (세 번째 변수는 이미 두 번 이상 리스트에 포함될 수 있음)
columns_to_drop = list(set(columns_to_drop))

# 결과 출력
print(f"상관계수 0.9 이상인 변수 쌍 개수: {len(high_corr_pairs)}")
print("상관계수가 0.9 이상인 변수 쌍:")
for pair in high_corr_pairs:
    print(f"{pair[0]} - {pair[1]}: {pair[2]:.2f}")

# 삭제할 칼럼들 출력
print(f"\n삭제 고려 할 칼럼 개수: {len(columns_to_drop)}")
print("삭제 고려 할  칼럼들:", columns_to_drop)

# 17.3s

from statsmodels.stats.outliers_influence import variance_inflation_factor
from statsmodels.tools.tools import add_constant

# 상수 항 추가 (회귀 분석을 위한 절차)
X = add_constant(df_cleaned)

# VIF 계산
vif_data = pd.DataFrame()
vif_data["Variable"] = X.columns
vif_data["VIF"] = [variance_inflation_factor(X.values, i) for i in range(X.shape[1])]

# VIF가 5 이상인 변수
high_vif_columns = vif_data[vif_data["VIF"] > 5]["Variable"]

print(high_vif_columns)

# 6m 40.2s

from sklearn.decomposition import PCA

X = df_cleaned.drop(columns=['Label'])

# 2개의 주성분을 사용하는 PCA 모델 생성
pca_2_model = PCA(n_components=2)
pca_2_model.fit(X)
pca_2 = pca_2_model.fit_transform(X)

# 3개의 주성분을 사용하는 PCA 모델 생성
pca_3_model = PCA(n_components=3)
pca_3_model.fit(X)
pca_3 = pca_3_model.fit_transform(X)

# 28.1s

# 로딩 스코어 (2개 주성분)
loading_scores_2 = pd.DataFrame(pca_2_model.components_.T, columns=['PC1', 'PC2'], index=X.columns)
print("로딩 스코어 (2개 주성분):")
print(loading_scores_2)

# 로딩 스코어를 (3개 주성분)
loading_scores_3 = pd.DataFrame(pca_3_model.components_.T, columns=['PC1', 'PC2', 'PC3'], index=X.columns)
print("로딩 스코어 (3개 주성분):")
print(loading_scores_3)

# 0.0s

least_contributors_pc1 = loading_scores_2['PC1'].abs().nsmallest(int(len(loading_scores_2) * 0.2)).index
least_contributors_pc2 = loading_scores_2['PC2'].abs().nsmallest(int(len(loading_scores_2) * 0.2)).index

# PC1과 PC2에서 공통으로 낮은 기여도를 가지는 칼럼
least_selected_columns = list(set(least_contributors_pc1).intersection(set(least_contributors_pc2)))
print("제거 가능성이 높은 칼럼들:")
print(least_selected_columns)

# 로딩 스코어 출력
loading_scores_least_selected = loading_scores_2.loc[least_selected_columns]
print("선택된 칼럼들의 로딩 스코어:")
print(loading_scores_least_selected)

# 0.0s

least_contributors_pc1 = loading_scores_3['PC1'].abs().nsmallest(int(len(loading_scores_3) * 0.2)).index
least_contributors_pc2 = loading_scores_3['PC2'].abs().nsmallest(int(len(loading_scores_3) * 0.2)).index
least_contributors_pc3 = loading_scores_3['PC3'].abs().nsmallest(int(len(loading_scores_3) * 0.2)).index

# PC1과 PC2가 PC3 에서 공통으로 낮은 기여도를 가지는 칼럼
least_selected_columns = list(set(least_contributors_pc1).intersection(set(least_contributors_pc2),set(least_contributors_pc3)))
print("제거 가능성이 높은 칼럼들:")
print(least_selected_columns)

# 로딩 스코어 출력
loading_scores_least_selected = loading_scores_3.loc[least_selected_columns]
print("선택된 칼럼들의 로딩 스코어:")
print(loading_scores_least_selected)

# 0.0s

# PCA 모델 정의, 전체 분산의 99.9%를 유지
pca = PCA(n_components=0.999)
# PCA 적용 및 변환
X_pca_999 = pca.fit_transform(X)

# 사용된 주성분의 수
print("사용된 주성분의 수:", pca.n_components_)

# 각 주성분의 설명된 분산 비율
print("주성분별 설명된 분산 비율:", pca.explained_variance_ratio_)

pca = PCA(n_components=10)
X_pca_10 = pca.fit_transform(X)

# 사용된 주성분의 수
print("사용된 주성분의 수:", pca.n_components_)

# 각 주성분의 설명된 분산 비율
print("주성분별 설명된 분산 비율:", pca.explained_variance_ratio_)

# 전체 설명된 분산 비율
print("전체 설명된 분산 비율:", pca.explained_variance_ratio_.sum())
