==> notebooks/07_data_preprocessing.ipynb <==
import os
import pandas as pd

# 모든 열이 출력되도록 설정
pd.set_option('display.max_columns', None)

output_dir = os.path.abspath("../data/processed")
os.makedirs(output_dir, exist_ok=True)

df = pd.read_csv("../data/processed/db_network_data_table.csv")

# 11.6s

# DB에서 받은 파일이 없을 시, csv 파일을 읽어와 하나로 합치고 칼럼명 좌우 공백 제거 및 inf 값 0으로 대체

# import os
# import pandas as pd
# import numpy as np

# # 모든 열이 출력되도록 설정
# pd.set_option('display.max_columns', None)

# # CSV 파일이 있는 디렉토리 경로
# csv_file_path = '../data/raw'

# # 디렉토리 내 모든 CSV 파일 리스트 가져오기
# csv_files = [f for f in os.listdir(csv_file_path) if f.endswith('.csv')]

# # 모든 CSV 파일을 읽어와 하나의 데이터프레임으로 결합
# df_list = []
# for file in csv_files:
#     file_path = os.path.join(csv_file_path, file)
#     df = pd.read_csv(file_path)
#     df_list.append(df)

# # 모든 데이터프레임을 하나로 합침
# df = pd.concat(df_list, ignore_index=True)

# # 컬럼명 좌우 공백 제거
# df.columns = df.columns.str.strip()

# # inf 값을 0으로 대체
# df.replace([np.inf, -np.inf], 0, inplace=True)

# 불필요한 칼럼
columns_to_drop = ['id', 'source'] 

# 단일 값으로 이루어진 칼럼
columns_to_drop.extend(['Bwd PSH Flags', 'Bwd URG Flags', 'Fwd Avg Bytes/Bulk', 
                        'Fwd Avg Packets/Bulk', 'Fwd Avg Bulk Rate', 'Bwd Avg Bytes/Bulk', 
                        'Bwd Avg Packets/Bulk', 'Bwd Avg Bulk Rate'])

# 중복된 칼럼
columns_to_drop.extend(['Subflow Fwd Packets', 'Subflow Bwd Packets', 'Avg Fwd Segment Size', 
                        'Avg Bwd Segment Size', 'SYN Flag Count', 'CWE Flag Count', 'Fwd Header Length.1'])

# 칼럼 제거
df_cleaned = df.drop(columns=columns_to_drop)

# 0.2s

# 음수값 처리: 음수값이 포함된 칼럼의 'Label' 값이 0인 데이터만 삭제합니다.

# 칼럼에 포함되어있는 모든 음수 데이터의 'Label' 값이 0 하나인 칼럼
negative_columns_with_only_zero_label = ['Flow Duration', 'Flow Bytes/s', 'Flow Packets/s', 
                                         'Flow IAT Mean', 'Flow IAT Max', 'Fwd Header Length', 
                                         'Bwd Header Length', 'min_seg_size_forward'] #Fwd Header Length.1 는 Fwd Header Length와 동일 칼럼으로 삭제

# 해당 칼럼의 음수 값이 있는 행 삭제
for col in negative_columns_with_only_zero_label:
    df_cleaned = df_cleaned[df_cleaned[col] >= 0] 

# 2.7s

# 결과 파일로 저장
# 최종 데이터를 'data_cleaned.csv'로 저장
df_cleaned.to_csv(os.path.join(output_dir, "data_cleaned.csv"), index=False)

# 완료 메시지 출력
print(f"\n데이터가 성공적으로 'data_cleaned.csv'로 저장되었습니다.")

# 58.3
